+++ etl/scripts/header.py
@@ -3,6 +3,7 @@ import logging
 import os
 
 from pe_memberdna.lib.job_manager import JobManager
+from pe_memberdna.etl.lib.s3 import input_data_validator
 from pe_memberdna.etl.lib.utility import *
 
 
@@ -13,8 +14,8 @@ def main(job, data_paths, config_validation):
 
     cast_sql = """
     select
-         cast(PURCH_HDR_ID as int) as PURCH_HDR_ID
-        ,cast(MBRSHP_SID as int) as MBRSHP_SID
+         cast(PURCH_HDR_ID as long) as PURCH_HDR_ID
+        ,cast(MBRSHP_SID as long) as MBRSHP_SID
         ,cast(SITE_NBR as int) as SITE_NBR
         ,cast(PURCH_DT as date) as PURCH_DT
         ,SALES_CHANNEL_ID
